make HTTP Content-Disposition testcases data driven
This also adds some new tests from http://greenbytes.de/tech/tc2231/ and
activates some that were still commented out "because of missing
implementation" even if the implementation is already there.

One of the new tests showed that I was accidentially converting all values to
lowercase because of a copy&paste error, this is fixed now.