Stop the build if an installed KDE4 is found which is not being overwritten.
We're having problems where people are able to compile the frameworks
because they are able to find headers which have been moved in the
frameworks branch. So it compiles for the developer and they push,
but then myself and Volker can't build it.

If the check is wrong, feel free to fix it. This is the first attempt.