kshareddatacache: Consolidate error handling.
This commit consolidates the error handling that goes on with various
unrecoverable failures to attach to the shared cache into a single
function which can always be run to reset KSharedDataCache::Private into
a state where mapSharedMemory() can be called again.

This minor reorganization is required to support an upcoming bugfix, but
also fixes the m_lock not being cleared even though the shared memory it
belongs to has been unmapped.