kshareddatacache: Call KSDCLock::initialize()
Till Adam pointed out in a review of Mac OS X changes to
KSharedDataCache that m_lock->initialize() is not called when it should
be. This happened to work on Linux and other POSIX platforms, but does
not work on the lock Till is working on.

(cherry picked from commit e851482f8333c871f808e2f0de076fcd210a3da3)