kshareddatacache: Don't lock if not mapped to shm.
If we're not mapped to shared memory then m_lock will be null. It is
unfortunately possible to be in this degenerate state however briefly,
and currently no part of the lock chain checks for this.
(cherry picked from commit bc3fa5c771b81daf85b5847cc5f751c968466c14)