Fix kbuildsycoca4 crash when KDE_NO_DEPRECATED is set
(e.g. in the Mobile profile).

This is because KDE_NO_DEPRECATED is only defined in kdemacros.h, which not
all files include anymore (now that it's not part of kdecore_export.h)

TODO: one "deprecated" define per lib, and generating the #define optionally,
as part of generate_export_header. Seems sensible?
CCMAIL: steveire@gmail.com