Support GnuPG2 in KNewstuff3
So far knewstuff has "gpg" hardcoded as the executable name for GnuPG. This
change also allows the use of "gpg2" instead, if the "gpg" executable is
not present (both are compatible as far as we are concerned here). Since
kdepim* requires GnuPG2, you are most likely using gpg2 already anyway, via
backward compatibility forwards. On MeeGo however this is not available and
packages for GnuPG1 and 2 are not installable at the same time.

REVIEW: 102439