Add KCompressionDevice
KCompressionDevice is a new class, with most of the current KFilterDev
code, but an improved API.
No use of KMimeType or mimetype strings in that code, the enum is used
instead.
The handling of None is new, it has to be added to the code.

For compatibility and for a slightly higher-level class which can
auto-detect mimetypes, we keep KFilterDev

Add a new filter NoneFilter for uncompress data
Add test for the NoneFilter