Fix KDirWatch when using Kernel 3.0
To know INotify is available in the kernel we are checking for a
kernel newer than 2.6.14, that's all allright but the problem is
that the Kernel version format has changed  and now we can't be sure
that it is going to be formed by 3 numbers.
Basically we where checking for:

%d.%d.%d and now it can be %d.%d as it is with 3.0

This patch what does is check if the kernel is 2.6, if it is then
it proceed with the version checking.

CCMAIL: dfaure@kde.org