kcharselect: Improved precondition checking.
Apparently bug 235020 (crash in KCharSelect) is still around even in
kdelibs 4.7.0.

Harald Sitter had noted in comment 46 to that bug that the QModelIndex
should be checked before it is used, and the int type of pos means it
could overflow to be negative. Neither condition is checked for yet, so
check for them now.

Hopefully we really fix bug 235020 (or at least reduce severity away
from being a crasher).

This commit is for what will become kdelibs 5.

CCBUG:235020.