Don't use "method static" (ahem) variables for per-instance data.
When I wrote that code I didn't know that static data in member functions
is "global" - all instances use the same data. So if there is more than
one KExtendableItemDelegate in an application, it makes Bad Things happen.
At least it is binary compatible to remove the bogus variables.

This is officially a brown paper bag bug.

Hopefully fixes
BUG: 259333