Fix a regression caused by a patch that addressed bug# 60898 some 8 years ago.
The client should handle the message body returned by the server as a result of
a PUT request since the only difference between a PUT and POST action is how the
Request-URI is supposed to be handled by the server. See RFC 2616 section 9.6.

More importantly this patch should fix the annoying and intermittent "server error"
message that pops up when browsing git.reviewboard.kde.org with Konqueror.