Fix performance regression in KFilePreviewGenerator
When switching from the details-view to the icons-view in Dolphin/
Konqueror a freeze might occur for several seconds. The root cause
for this is commit 41854146040f8638fffb13b20a75f6ecda2837a6. Using
the DateChangeObtainer only worked for the icons-view but in case
of the details-view still inside Qt a layout change got triggered.
As we know that in our case changing the icon won't result in a
change of the layout the signals of the directory-model will get
blocked temporary. By using the new view-engine such ugly workarounds
won't be needed anymore but KFilePreviewGenerator is still used
by other applications.

Thanks to David Faure for investigating into the root-cause!

BUG: 278899
FIXED-IN: 4.7.3