When querying for audio files only use the mimetype.
We sadly cannot depend on the nfo:Audio type. That would be the fastest
solution but Strigi still fails to index a lot of audio files. Thus we
depend on the very reliably mimetype from KMimeType.
However, if we do that we already get all the audio files. Thus, there
is no point in adding the additional ResourceTypeTerm which only makes
the query that much slower.

BUG: 276444
FIXED-IN: 4.7.3