Handle kext:altUrl in Query::Query include and exclude folders.
The new symlink handling adds symlinks as kext:altUrl properties to the
target files. Thus, whenever we search in a symlink folder we need to
restrict by kext:altUrl instead of nie:url.
TODO: There is still one case not handled though: imagine we are
search recursive in one folder which contains symlinks to other folders.
Then the contents of those other folders are not searched since their
altUrls are not taken into account. This, however, would mean a
considerable overhead which I am not sure how to handle yet.