Improved way to handle kext:altUrl.
This is a port from nepomuk-core which I push as a new commit since I
cannot rebase the existing branch I already pushed.

Now that we store symbolic links in kext:altUrl the easiest way to
search symlinks is to simply always matchin both nie:url and kext:altUrl
whenever a client asks for nie:url.