Import git@gitorious.org:qmime/qmime.git temporarily, much like libinqt5.
This required porting it to CMake, and fixing compilation with our stricter
flags. Please don't make changes to this code here, upstream development
is in qmime.git (and will soon require CLA).

Warning: some test files were converted with dos2unix to workaround the
git hook refusing them. Had to add a hack in test.url to make it work.