Do soley rely on the mimetype of video files rather than the rdf type.
Using the combination of both results in search result duplicates.
Using only the rdf type is a problem since libstreamanalyzer fails to
add that type on many video types. While this is now "fixed" by adding
nfo:Video as a type to all files which contain "video" in their mimetype
legacy data is still affected.