Use KCoreConfigSkeleton argument type where possible inside the KConfigDialog
Use case: there are applications, like kanagram, which would be nice to have
running on several platforms, as in handsets; for instance Harmattan on N9. It
would be nice to use the same settings code generation in certain cases for all
the platforms. The additions of KConfigSkeleton on the top of
KCoreConfigSkeleton are the font and color settings which are currently not
used in couple of KDE applications. Hence, it should not be mandatory. The
kdeui module is unlikely welcome on mobile platforms, especially in appstores
with its sizes and complexity for no real need.

KConfigDialogManager has apparently already two constructors (ie.: the need
already arised previously for such an approach); one with KConfigSkeleton
argument type, and yet another with KCoreConfigSkeleton. It looks like a
situation where the KCoreConfigSkeleton version was added for good.

The KConfigDialog constructor does not handle KCoreConfigSkeleton argument
type yet; it has probably somehow been missed so far. Changing the current
constructor to KCoreConfigSkeleton usage is a good way because it does not
cause any API break; a simple recompilation is enough in client applications.