Make the writeConfig() method a public slot
REVIEW: 103829

According to the kconfig-xt example [1], it seems to me we need to call this
method for settings changing. Therefore, it is not enough to just set the
certain properties separately. I currently need to wrap this writeConfig()
method in a helper class in order to get an invokable or slot version.

There are at least two ways of addressing the issue in question:
1) Make the current method Q_INVOKABLE
2) Make the method actually a slot

I personally chose the second alternative since the method already returns
void, thus a potential candidate for being a slot. According to the Qt
documentation, it is not a problem for having a virtual (public) slot:
"You can also define slots to be virtual, which we have found quite useful in
practice."

[1] http://techbase.kde.org/Development/Tutorials/Using_KConfig_XT#Example

Tested on Archlinux for building only.