Allow protocol-specific ThumbCreator plugins to be used for folders.
Old-school ThumbCreators (those which work on local files) are
handled by the PreviewJob already. It passes their name down to the
thumbnail:/ KIO slave so the latter knows which plugin to use.
For remote protocols this is not the case. The PreviewJob only checks
if there is a plugin which can handle it but does not send the plugin
name along. This results in the thumbnail KIO slave to default to its
internal folder-preview algorithm when there is actually a specialized
ThumbCreator plugin available.
This patch fixes the situation by also passing along the names of remote
protocol plugin.
REVIEW: 103882