Replicate the existing KConfigDialog constructor and change one argument type
REVIEW: 103716

Use case: there are applications like kanagram which would be nice to have
running on several platforms, like handsets; for instance Harmattan on N9. It
would be nice to use the same settings code generation in certain cases for all
the platforms since the additions of KConfigSkeleton on the top of
KCoreConfigSkeleton are the font and color settings. These are currently not
used in many KDE applications. Hence, it should not be mandatory. The kdeui
module is unlikely welcome on mobile platforms, especially in appstores with
its sizes and complexity for no real need.

KConfigDialogManager has apparently already two constructors; one with
KConfigSkeleton argument type, and yet another with KCoreConfigSkeleton. It
looks like a situation where the KCoreConfigSkeleton version was added later.

KConfigDialog does not have a constructor yet with KCoreConfigSkeleton argument
type yet; it has probably somehow been missed so far. Changing the current
constructor to KCoreConfigSkeleton usage is not possible in the 4.X major
version because of the consequences (ABI breakage). Thereby, the freshly
replicated constructor. The proper fix can be filed against frameworks where
there is only one, and properly working constructor.