Fix Bug 247338 - Sonnet spellchecking dialogue: Cancel action not implemented properly
signal languageChanged was emitted all the time when a config was changed in widget.
So it was not as defined in api.
Created a new signal configChanged() better name and keep old signal.
Remove it for KDE5 and clean up this class
(We need to make some work on sonnet for kde 5)

FIXED-IN: 4.8.1
BUG: 247338