Create the ResourceWathcer when loading the data in Nepomuk::Resource
There is no point in updating the cache, if the cache hasn't been created.

In a lot of places (eg - queryservice ) Nepomuk::Resource::resourceUri() is used, which
lands up calling determineUri and would then send a dbus call for the resource
to get added to the ResourceWatcher. We don't want that additional overhead.

REVIEW: 104029