Fix normalization of selections by KSelectionProxyModel.
The previous code ran two iterators over a QList, and used erase()
on one of them. The problem was that QList invalidates all iterators
on it when erase is called, not just the ones following as this code
had assumed incorrectly.

The solution is to populate another QItemSelection instead of
modifying in-place.

CCMAIL: cosenal@gmail.com