kshareddatacache: Improve item finding/insertion.
Noted in a code review by Ralf Jung, KSharedDataCache suffered from a
couple of deficiencies in its handling of finding and inserting items:

1) When looking for entries in the cache, quadratic probing is used.
Assume an item foo is placed in the 3rd spot checked due to existing
items in the 1st and 2nd spot. If that 2nd item is later removed,
KSharedDataCache would never find items that probed logically after
that, as the search would stop once an item was checked with
useCount == 0.

The fix is to check until the item is found or the max probe count is
reached, unconditionally.

2) When inserting an item, it is possible for two entries with different
keys to have the same reduced hash. If this occurs then when inserting
the second item it would be placed after the existing one. Previously
KSharedDataCache would never find the second item afterwards as it would
match the first item (with identical hash), verify it didn't have the
right full key, and then stop checking at that point. It is very
unlikely for two items to have the same hash, so we now remove existing
items with the same hash code if this occurs. As it turns out the fix
for 1) will ensure we find existing 'masked' items even though they are
no longer permitted.

The work is Ralf's, rewritten a bit by myself to minimize changes to
stable kdelibs to the minimum necessary to correct the behavior. In fact
Ralf's analysis (on this and a couple other points) has been so good
that I've added him to the copyright header.

CCMAIL:faure@kde.org
CCMAIL:ralfjung-e@gmx.de