kauth-api-changes: Allow only asynchronous execution, remove ActionWatcher, and introduce the new ExecuteJob.
With these changes, now only async executions are allowed with a KJob-style asynchronous operation.
This relieves the pain from the backends to handle different methods (one can always suppose we're in an asynchronous realm),
and at the same time "hides" ActionReply from the client implementation.

Now all the relevant information for the client are advertised through the Job itself. ActionReply is still here to stay
since its serialization features are vital for the helper, and because we don't want to break the API too much.