kauth-unit-tests: Add a way for choosing which bus to use to DBusHelperProxy
This is not useful from KAuth perspective, but allows us to use the backend for unit testing
(which is a real win-win, since we also get to test backend code). By feeding it a session bus,
we can create a fake service inside the unit test itself without the need of registering
authorizations on the system bus, which require installing some files.