imload_png: Remove incorrect extra libpng call.
KHTML's PNG decoder has an apparently extraneous call to a libpng
function, as reported by libpng developer John Bowler.

png_read_update_info already does the work required by
png_read_start_image. This extra work causes the upcoming
libpng-1.6.0-beta18 to fail to load images at all (they will probably
downgrade this to a warning before the 1.6.0 release, but should still
be fixed).

Unfortunately I neglected to commit this when the patch was first
approved, so it will not be present in a release until 4.8.3.

BUG:296305
FIXED-IN:4.8.3