Clean up ktimezonedrc properly after each unit test
ktimezonestest_p.h already did try to clean up the config after testing,
but the folder to be removed was hardcoded (~/.kde-unit-test/share).
This does work for ktimezonestest, but not for ktimezonewidget_unittest.
That test uses the same private header, but does not change KDEHOME to
~/.kde-unit-test, such that the user's normal config was modified by the
test, and the changes persisted even after the test. This then caused
applications like, e.g., Dolphin to output lots of unwanted error
messages at runtime.

This commit changes that by determining the full path of ktimezonedrc at
runtime and deleting that file.
BUG:295383