Repair KAuthorized::authorizeUrlAction
(disabled in recent kconfig splitting, which broke kdirlister)
To solve the dependency on KProtocolInfo, kconfig now has a kcoreauthorized
with all the simple string manipulation, and KIO has the api that takes QUrls.

Since KAuthorized is a namespace, the split up could be done without affecting
source compatibility.
Yay for namespaces (compared to classes with static methods).