Use KTzfileTimeZone instead of KSystemTimeZone on UNIX
Use KTzfileTimeZone in preference to KSystemTimeZone on UNIX systems,
since use of the standard system libraries by KSystemTimeZone
requires the use of tzset() in several methods. That function reads
and parses the local system time zone definition file every time it is
called, and this has been observed to make applications hang for many
seconds when a large number of KSystemTimeZone calls are made in
succession.

BUG:281360