"data" protocol: remove broken shortcut for charset access
This shortcut always returned the default value for the charset initially set:
US-ASCII. We later parse the charset from the URL and set it in the attributes
map, but ignore it for our own parsing of the URL contents.

This can't really make anything worse: we always used the US-ASCII encoding
which is a subset of everything else so everything that properly worked until
now isn't expected to get broken.