Fix animation of scaled GIF images.
Firstly, implement the AnimProvider::clone function in the GIF loader.
This enables the creation of scaled AnimationProviders.

Secondly fix the ImagePainter so that it uses the set size instead of
the size of the base (unscaled) image.

Lastly fix RenderImage::updatePixmap so that it takes scale into
account when deciding which rect to repaint.