Never cache the "unknown" icon.
When reviewing this code years ago at the KSharedDataCache integration,
it actually kind of made sense to cache "unknown" icons, when there were
still so many unimplemented icons extant.

Now there's no reason to do that, unknown icons are rare and it is very
difficult to force the icon out of the cache after the issue is "fixed",
so don't allow them to be used if cached.