Do not force the largest-available icon size to be used.
The current icon-selection code "locks-in" the largest size icon already
found that is otherwise suitable. This is undesirable since downscaling
the icon which is the next step up in size should generally be quicker
and give a more size-appropriate representation on-screen.

Instead of skipping icons that are smaller than an icon already found
(even if they are a better "fit" for the desired size), we use a simpler
test to skip icons that require upscaling when we've already found an
icon that can be downscaled. Otherwise we prefer icons that are closer
to the desired size.

Issue noted and debugged by Ignat Semenov.

CCMAIL:ragnarokk91@gmail.com