tests: add a runner for the ECMAscript Test262 suite
Usage:

 cd /some/where
 hg clone http://hg.ecmascript.org/tests/test262/
 cd build/for/kdelibs
 cmake -D ECMATEST_BASEDIR=/some/where/test262 .
 make
 ctest -R ECMAscript262

Especially chapter 15 will take a while, so running with "-j 4" will likely
speed things up a bit (replace 4 with the number of your CPU cores).

To run a single test line:

 ECMATEST_BASEDIR=/some/where/test262/  ./kjs/tests/ecmatest runAllTests:7.6-10

I've set up a list of things from the different test chapters that I know KJS
is currently failing. This is the list of things that constantly fail on the
2 hosts I've tested this on. The results differ between those hosts and even
between subsequent test runs, but those listed will always fail.