kshareddatacache: Validate cache page size.
This commit ensures that the cache page size is actually a power-of-2
and within the band of possible sizes that could possibly have been set.

If this is not the case the cache is assumed corrupted and reset.

This should help with any cache-corruption bugs caused by a wrong cache
page size (although these don't exactly make themselves obvious). More
fixes to follow...

This one /should/ fix 274252 outright and may be of interest to several
others.

CCBUG:249362,253665,243573,281217,297815,293954,293447,270915,255233

BUG:274252
FIXED-IN:4.8.4