kshareddatacache: Sanity checking for intCeil.
intCeil is only ever used with unsigned int, so remove the template.
This allows for proper validation of the denominator, a corrupted-cache
exception will be thrown if intCeil is called incorrectly.

Passes the kshareddatacache and kiconloader_unittest tests, and my
limited fuzz testing of kpat and plasma themes.

Should fix bug 282027 and any duplicates mentioning "intCeil(0, 0)" in
the backtrace.

BUG:282027
FIXED-IN:4.8.4