KJS: Optimize iteration of empty hash tables
Reviewed by Geoffrey Garen.

Iteration of empty hash tables was showing up on simple page loading
profiles due to the use of hash tables in the
NodeListsNodeData::invalidateCaches and
NodeListsNodeData::invalidateCachesThatDependOnAttributes functions.
It's worth optimizing the case of an empty table.

* wtf/HashTable.h:
(WTF::HashTable::begin): Return the known-good end iterator when the
table is empty. This makes iterating an empty table faster because we
avoid skipping empty and deleted buckets.

https://bugs.webkit.org/show_bug.cgi?id=87215

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@118187 268f45cc-cd09-0410-ab3c-d52691b4dbfc