KDE4Macros.cmake: fix wrapper shell if target has a different name
We query the target LOCATION here to use it in the wrapper script as executable
name. The CMake documentation about this says:

   Do not set properties that affect the location of a target after
   reading this property.

But we need to do this e.g. for kjs_bin which should be called "kjs" but CMake
doesn't want to have a binary and a library with the same name.

For newer CMake versions we can get at least the kjs_bin.shell wrapper right by
using generator expressions for the binary name.

(cherry picked from commit a1004b4423eafb2fc6b04f30002f42c8ea29150f from KDE/4.8)