Fix KSharedDataCache build with KDE4_ENABLE_FINAL set.
Support level for "final" builds remains "meh" with the buildsystem
types, but this did work in kdelibs 4.8.3 so I've received consensus
support for applying a patch provided by the Mageia devs.

Thanks to Funda Wang from Mageia for the problem description and
proposed fix. The "fixed-in" may be incorrect based on whether we
reissue a kdelibs tarball for the Soprano issues but it should be close.

BUG:301419
FIXED-IN:4.8.5
CCMAIL:kde-packager@kde.org