kshareddatacache: Correct the findEmptyPages loop boundary.
I did review the last patch to remove 'pagesNeeded' from the loop test,
but I missed that it has '+ 1' instead of '- 1' (which is still too
conservative). The correct thing is to not have any addition at all
since we use operator<() for the test.

Graciously pointed out by kde_pepo, thanks!