Generate the moc files (add missing dependency)
Alex wrote:
The difference here is that qt4_generate_moc() from KDE calls
macro_add_file_dependency() to make the cpp file depend on the generated moc
file, so the moc is created. The one coming with cmake doesn't do this, so
this dependency is not there and the file is not created.

CCMAIL: kde-buildsystem@kde.org