ResourceManager: Do not double initalize the NepomukMainModel
The NepomukMainModel is initalized once when it is created and then the
second time in ResourceManager::init.

This results in a second connection being made to nepomukstorage, and an
extra thread being spawned over there. The old thread is eventually
deleted, but the whole process is rather pointless.

REVIEW: 105281

Copied from nepomuk-core 847d642f876507bda220750468afbba3adbccbfa