Avoid automatic clearing of the ResourceManager cache
The current limit was hard coded and wasn't based on any real world
usage. With the new dolphin features and amarok being ported to Nepomuk,
it is very common to have more than 10k resources.

When one overcomes that number, then for each new Resource the entire
kickoff list is searched in an attempt to find ResourceData* which are
no longer being used.

Instead, now whenever a ResourceData* is no longer being used, it is
simply deleted.

REVIEW: 105516