Turn smb: into smb://, remote: into remote://, in prettyUrl.
It's prettier not because we like slashes, but because it makes it easier
to type a hostname afterwards. As requested by Alex Fiestas.

On the other hand it also turns applications: into applications:// which
doesn't hurt much, but doesn't make much sense either (no hosts there).

In Qt5 we can fix this properly, since QUrl differenciates empty-host
and null-host, finally. So we can do the above redirections in the
appropriate kioslaves instead.