Resouce: Properly implement genericIcon
The old generic icon just used to look at the string present in
nao:hasSymbol. However, that was incorrect as nao:hasSymbol has a range
of nao:Symbol, and not of a string.

This patch results in proper icons being shown in the Nepomuk KRunner,
as it uses genericIcon to get the icon name.

REVIEW: 105745