Make KModifierKeyInfoProvider inherit QObject, rather than QWidget
This makes crashes like the one I'm closing with this report impossible,
even though the root cause might be somehing else, possibly related to
Qt and particular settings of certain styles.

In any case, that class does not need anything from QWidget any more, so
this change would make sense even without the crash.

Thanks to Aurelién Gâteau and David Faure, both for coming up with the
idea to fix it this way in the first place and for explaining why
deriving the class from QWidget did make sense in the past.

BUG: 303169
FIXED-IN: 4.9.1
REVIEW: 105920