Do not pass -lc to the compiler on BSD systems.
It is not completely clear why this flag was being specified in the
first place: it was added in the commit that created
FindKDE4.cmake (KDE SVN r497283), and seems to have been ported
automatically from the autotools code.

In autotools, its origins can be tracked back to KDE SVN r141814,
which was a libtool update. In that case, -lc was only passed on AIX
systems anyway.

Cherry-picked from commit 7752d6d5b9b67987d4631cca4ce192069697a4e2 to
extra-cmake-modules.