Fix unittests that need to call kbuildsycoca.
In KDE4 they would pass env vars (e.g. KDEHOME etc.) pointing to ~/.kde-unit-test,
but in KF5 QStandardDirs is switched to "test mode" without the use of env vars.
This means passing "we are in test mode" to kbuildsycoca (so that it generates
ksycoca in the expected place) now requires a command-line argument.