Ensure slotChanged is called in all instances
Move slotChanged call to the broadcastDone slot
so that it is called in all instances of the device; in fact broadcastDone
is called by dbus.  This makes sure that the device state is correctly updated
when modified by some out-of-process routine, e.g. when mounted via
kde-runtime/soliduiserver actions.

This *should* fix bug 268020, please check
CCBUG: 268020