Match font style with human-readable style name using custom identifier.
Style strings returned by QFontDatabase::styleString() are sometimes
not surviving serialization/deserialization roundtrip, which can cause
wrong style to be highlighted in style box when the dialog is opened.
If the user then closes the dialog without touching the style box,
the resulting font style is changed. To avoid this, construct
internal style identifiers and use them for matching instead.

Solution provided by Casper Boemann.

BUG: 305747
FIXED-IN: 4.9.1