Move kde plugins from %lib/kde5 to %lib/plugins/kf5
And qt plugins from %lib/kde5/plugins to %lib/plugins.
This way we have all plugins in the same dir, and that dir conforms
to the xdg discussion last year. It will also make it easier to use
the same framework for loading both, later on.

However this means a SIC in cmake vars: installing qt plugins must be
done with QT_PLUGIN_INSTALL_DIR rather than PLUGIN_INSTALL_DIR/plugins
(which, on hindsight, was a nasty hardcoded path...).