kshareddatacache: Add simple fallback spinlock.
This is a very simple spinlock busy-waiting class to allow for a
process-shared, shared-memory lock when the operating system does not
provide a suitable anonymous lock.

Based on code from Till Adam and Thiago Macieira.

One of the intended targets is OpenBSD which doesn't seem to support
sched_yield, so they get nanosleep instead. This check should be shifted
to a CMake check once that patch is integrated.

Since this spinlock "should work everywhere" I remove support for
finding thread-only locks (to be clear, support for such has always been
a bug itself). If no standard POSIX primitive is available for
process-shared use we fall back to this spinlock.

Code was tested by artificially forcing the spinlock to be used,
removing all caches, and then running the test case and several heavy
users of the class (e.g. KPat and lskat).

This should hopefully be sufficient to eliminate the crashing
experienced in OpenBSD although I intend to continue merging the CMake
cleanups (but in KDE/4.10).

BUG:306192
FIXED-IN:4.9.2