kjs: FunctionObject prototype attribute should be [[Writable]]: true, [[Enumerable]]: false, [[Configurable]]: false ECMA Edition 5.1r6 - 15.3.5.2
the prototype must have DontEnum property otherwise it shows up as enumerable key.

In more detail, this fixes the following prototypes for created
1) function foo() {};
2) var x = function foo() {};
3) var x = new Function();

for the 1. I need the change in nodes.cpp
for the 2. I need the change in bytecode/codes.def
for the 3. I need the change in function_object.cpp

REVIEW:105122