kjs ecma262 test: get the pass/fail counting more meaningful
Count in terms of the ECMA262 testsuite, i.e. if we fail a test we know to fail
this counts as passed for QTest, but it is a fail in the testsuite. This way
we will end up with meaningful numbers we can compare with what we get when
running the testsuite in the browser.