Fix infinite recursion in kdeui tests.
QCoreApplication::installTranslator cannot be called from the KGlobalLocaleStatic
constructor, because the LanguageChange event calls tr(), which calls KLocale::global(),
so KLocale::global() must exist already (i.e. KGlobalLocaleStatic ctor must be finished).

So move it to a separate init() method, but this requires a recursive mutex and
a bool. I wonder if there's a simpler solution...
CCMAIL: thiago@kde.org