Changed the global "Treat all cookies as session cookies" option to a per domain/cookie policy setting called "AcceptForSession".
As the name implies the policy allows a user to accept a given cookie that would
automatically expire it at the end of the current session regardless of the
cookies' expiration date.

The patch has a corresponding patch that takes care of the cookie configuration
module and was provided by Regorio Guidi.

CCMAIL:gregorio.guidi@gmail.com
CCBUG:54300
REVIEW:106618