Start of async package installation and deinstallation
We basically move two sync methods (installPackage and uninstallPackage)
into it' QThread-backed KJob class, and handle installation this way.
This allows us to not block the UI, since package installation,
involving I/O can take a long time.

Tests are broken, so is the code right now. It's just becoming a little
bit to big to not have it VCS'ed and shared.
	plasma/.kdev4/