Support Python 3.2's new method of handling byte compiled files in PythonMacros.cmake.
Python 3.2 changed the compile mechanism by creating a __pycache__
directory. This obviously breaks a lot of stuff that assumed that pyc
files were stored in the same directory as the py files. This change
adds a check, and adjusts files only in case Python 3.2 is found. Notice
that currently I have way of testing if the suffix (*cpython) is the
same on all supported OSes.

Also, I fixed a number of typos that made the macro work by chance all
these years.

REVIEW: 106707