Remove KDE_NO_PHONON hackery
Previously the Phonon headers and library was always required to build,
but when building with Qt5 the build system made sure never to link
against the library. Thus the build would work (with features that
require Phonon) even if the only available phonon was built with Qt4.
(Which previously was the only possibility, as Phonon didn't build with
Qt5).

With this patch, the Phonon headers and library is optional to build,
but if they are found they will be used. Thus building with Qt5 and a
Phonon built with Qt4 will no longer work. However, this patch does
successfully enable the missing features, provided Phonon was built
with Qt5.

In other words, these combinations are now supported (* are new with this patch):
Qt4, No Phonon*
Qt4, Phonon 4 on Qt4
Qt4, Phonon 5 on Qt4
Qt5, No Phonon*
Qt5, Phonon 5 on Qt5*

While these combinations are not supported (* previously (sort of) worked):
Qt4, Phonon 4 on Qt5
Qt4, Phonon 5 on Qt5
Qt5, Phonon 4 on Qt4*
Qt5, Phonon 5 on Qt4*
Qt5, Phonon 4 on Qt5

Note: "Qt5, Phonon 4 on Qt5" is fine as far as kdelibs is concerned, but
Phonon 4 won't actually build on Qt5, making it a no-go.

REVIEW: 106987