kdecore/date: Use consistent include guard for TZ cache testing.
There is a feature to test the KDateTime caches, which adds some
KDE_EXPORT'ed variables when compiled with a certain #define. This makes
it impossible to build just kdatetimetest without building kdelibs with
COMPILING_TESTS (which I can't find in any of the cmake files).

The fix is just to consistently enable the cache tests only when kdecore
is compiled using the proper compiler defines, by using the same
defines.