kcalendarsystem: Remove the deprecated weekDayOfPray() method.
It really don't belong in the calendar, and removing it lets us remove
some very bogous values for calendars not tied to an Abrahamitic religion.

Note that API remains as an depecated alias for KLocale::weekDayOfPray(),
which may return 0 for no cultural week day of prayer, which the existing
documentation allows, but no calendar system ever did.