Make iconName() virtual so that it can be determined on demand.
Right now kuriikwsfilter.cpp loads all the icons for all the search
providers upfront (see the call to setIconName), leading to lots of
dbus calls to the favicon kded module, even when simply selecting text
in konsole... It would be much more efficient to do this on demand
in KUriFilterData::iconNameForPreferredSearchProvider, i.e. only if a match
was found and only if the app cares for the icon.

Not sure how to fix this in KDE-4.x though.
CCMAIL: adawit@kde.org