Consider two KFileItems 'equal' if they share the same URL
Up to now, KFileItem::operator==(const KFileItem&) only returned true if
both items shared the same d pointer. However, the implicitly shared
d pointer is an implementation detail of the class, and making the
comparison depend on it might not be expected by the developer using the
class. Fixes a problem in Dolphin, where the unexpected outcome of the
comparison caused permanent high CPU usage.

CCBUG: 304986